requirements.txt
pygame
struct
PyOpenGL
Pillow